% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/peb-functions.r
\name{PEB.conf.int}
\alias{PEB.conf.int}
\title{Confidence intervals for personalized PEB thresholds}
\usage{
PEB.conf.int(p, n, conf.level, params)
}
\arguments{
\item{p}{probability (specificity)}

\item{n}{Number of previous observations in the individual's history. If \code{length(n) == 1}, n is expanded to \code{seq(0,n)}.}

\item{conf.level}{Confidence level of the interval.}

\item{params}{A PEBparams object returned by \code{PEBparams(...,method='iterative')}}
}
\description{
Confidence intervals for personalized PEB thresholds.
}
\examples{
PEB.conf.int(p=0.95, n=6, conf.level=0.9, params=params)
}
\seealso{
Other peb: \code{\link{PEBparams}};
  \code{\link{pebVarArgs}}; \code{\link{peb}},
  \code{\link{ppeb}}, \code{\link{qpeb}},
  \code{\link{zpeb}}
}

