% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/peb-functions.r
\name{peb}
\alias{peb}
\alias{ppeb}
\alias{qpeb}
\alias{zpeb}
\title{Functions that apply the Parametric Empirical Bayes (PEB) algorithm}
\usage{
zpeb(x, ..., [n, ybar,] na.rm = FALSE, details = FALSE)

ppeb(x, ..., [n, ybar,] na.rm = FALSE, details = FALSE)

qpeb(p, ..., n, ybar, verbose = FALSE)
}
\arguments{
\item{x}{A vector of marker values to be interpreted by the PEB.  \strong{Note}
If n and ybar are not specified, x is interpreted to be a series of observed
marker levels from a single individual}

\item{n}{\code{n[i]} is the number of prior observations from the same
individual which preceed the i'th observation (\code{x[i]}).  Note that
n and ybar are required paremeters for qpeb and for zpeb and ppeb, must
both be present or both be absent.}

\item{ybar}{\code{n[i]} is the mean of prior observations from the same
individual which preceed the i'th observation (\code{x[i]}).  Note that
n and ybar are required paremeters for qpeb and for zpeb and ppeb, must
both be present or both be absent.}

\item{na.rm}{If TRUE and the parameters \code{n} and \code{ybar} are not passed, then
mising values in X are ignored when calculationg n and ybar from x. \strong{Note}
that if \code{n} and \code{ybar} are not passed and all values following the first missing
value in x will also be missing.}

\item{details}{If True, a data frame containing resutls from each
sub-calculation is returned,otherwise a vector with the z-scores
or p-values is returned. (Default = FALSE)}

\item{conf.level}{(optional) confidence level for the False Positve Rate (FPR)
for the thresholds provided \code{qpeb()}. If conf.level is specified, \code{length(p)}
must be 1 and the PEB parameters must be specified by the object returned by
\code{PEBparams()} using the \code{'Iterative'} method, as in:

\code{params = PEBparams(x,id,method='Iterative',iterations=2);
qpeb(p,params,conf.level=0.95)}}
}
\description{
Functions that apply the Parametric Empirical Bayes (PEB) algorithm.
}
\details{
\code{zpeb(x,[n,ybar,]...)} returnes the z-score for a marker value after applying the PEB algorithm

\code{ppeb(x,[n,ybar,]...)} returnes the p-values for a marker value after applying the PEB algorithm

\code{qpeb(p,n,ybar,...,conf.level)} returnes the quantiles for a marker if the PEB is
to algorithm with a specificity \code{p}. In particular, if the normality
assumptions are satisified, the probability that a marker exceeds
\code{threshold = qpeb(p,n,ybar,...)} is \code{p}.
}
\note{
Note that each function requires just 2 of the 4 parameters v,sigma2,tau2, and beta (=beta1)
\strong{Note} also that mu,sigma2,tau2,v, and beta may be substitured by a the object returned by \code{\link{PEBparams}}
}
\examples{
n = 10
m = 200
params <- PEBparams(x = rnorm(m*n) + rep(rnorm(m),each=n), id=gl(m,n,m*n))
x <- rnorm(10) + 6
plot(seq(length(x)),ppeb(x,params))

# interesting facts about the PEB...
b1 = .7

# this should be uniformly distributed by definition
hist(ppeb(rnorm(5000)*sqrt(1-b1),mu=0,v=1,beta=b1))

# this should be positive on the outset.
hist(ppeb(rnorm(5)*sqrt(1-b1) + 3,mu=0,v=1,beta=b1)) # 10/sqrt(1-b1) standard deviations above normal!!

# but even this should be uniformly distributed in the long run
hist(ppeb(rnorm(5000)*sqrt(1-b1) + 10,mu=0,v=1,beta=b1)) # 10/sqrt(1-b1) standard deviations above normal!!
}
\seealso{
Other peb: \code{\link{PEB.conf.int}};
  \code{\link{PEBparams}}; \code{\link{pebVarArgs}}
}

