% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/peb-utils.r
\name{pebVarArgs}
\alias{pebVarArgs}
\title{Calculate each of the variability estimates used by the PEB}
\usage{
pebVarArgs(...)
}
\arguments{
\item{mu}{The population mean}

\item{sigma2}{The within group variation}

\item{tau2}{The between group variation}

\item{v}{The total variation (=sigma2 + tau2)}

\item{beta1,beta}{The proportion of toal variation attributable to between group variation (=tau2 / v)}
}
\value{
a list containing the values sigma2, tau2, v, beta1, and mu,
and a function 'bn' which calculates bn for various values of n
given these parameters in this
}
\description{
A utility function used to calculate each of the variability
estimates used by the PEB algorithm from from any two of them
}
\note{
mu,sigma2,tau2,v, and beta may be substitured by a the object returned by \code{\link{PEBparams}}
}
\seealso{
Other peb: \code{\link{PEB.conf.int}};
  \code{\link{PEBparams}}; \code{\link{peb}},
  \code{\link{ppeb}}, \code{\link{qpeb}},
  \code{\link{zpeb}}
}

